Add WiFi communication and exchange of positions
When pressing the B (middle) button within the first three seconds of a
reset the M5Stack will act as a access point (AP) and will listen to UDP
messages. As soon as it received a message it will send its x- and
y-position back to that IP. The UDP server is currently hosted on the
default IP 192.168.4.1 and uses port 2390 (see serverIP and serverPort).
After the three seconds it will try to connect as a client and search
for the hosted WiFi (see AP_SSID and AP_PASS).

There is already a bearing calculated from the x- and y- positions but
for now there is now lift off state sent or received.

Another TODO would be to refactor and split up the pmw_mouse_sensor code
into Sensor, M5Stack, WiFi and App-Control / Main-Loop concerns (or
something like it).