Add first version of PDF generator
It can dynamically create PDFs with a pattern image as input.
The dynamic part looks at sensor resolution and capture area to find the
pattern pixel DPI the sensor should be able to read. It then uses this
DPI to scale the image accordingly when placing it on the PDF.