Add data folder for SPIFFS uploads
This folder will hold the data which is uploaded as SPIFFS to the
M5Stack and then stored on the SD card by the M5Stack-SD-Updater
library . Which of course has to be loaded onto the device to begin
with.
By this mechanism the SD-Updater menu can be updated without the
manually taking out the SD card.
To upload run:
`platformio run --target uploadfs`