Add the IMG_CAPTURE app
This app just captures the raw image, draws it to the display and sends
it over serial. If any button is pressed it will send a header byte to
indicate that a frame should be analysed.
It is used for an accompanying (python) script on the PC that receives
the image data. This way the image analysis can be done on the more
powerful PC hardware.