Add functions for access with dottrack_stream.py
These new functions will allow dottrack_stream.py to access and use the
image analysis pipeline.
This commit also fixes some issues with the dpi value set on the
captured image. Once the captured image is created it will set the
dpi to the cameras dpi and not the de Bruijn torus dpi. Minor fixes in
subsequent functions were needed.