Add Shius decoding algorithm for de Bruijn tori
The new decoding algorithm[1] should be much faster (~20x) than the old
brute force one.
The algorithm can (recursively) refer the brute force look up to the much
smaller source array the big one was constructed out of.
For now the implementation only works for the 256x256 4x4 de Bruijn
torus and does not use recursion yet. It also only implements the
decoding for arrays of type 1 because the construction of the 256x256
4x4 de Bruijn torus seems to only use type 1 constructions.

[1] Shiu, W.-C. (o. J.). Decoding de Bruijn arrays as constructed by Fan
et al.