Fix minor issue and clean up PDF creation script
It will set the width/height of the cropped image according to the pixel
width/height and not the overall printable areas width/height. In
reality this will only improve the accuracy very slightly because it
only varies within a single pixels width/height.
The clean up simplifies the code a bit and changes the interface a
little. It also adds a bit of documentation.