Fix memory issue in the get_test_frame function
The resizing/upscaling in the get_test_frame function was applied to the
whole de Bruijn torus image which led to extreme memory usage when using
bigger de Bruijn tori. For example did the (4096, 8192; 5, 5) de Bruijn
torus at 150 dpi take up around ~8.7GB memory.
The solution is to crop the original image down to the relevant area
(where the camera sits) and scale only that portion.