Add first auto capture script (only movement)
This first version of the auto capture script implements the movement of
the M5Stack by the AxiDraw across the page.
It can also read images from the serial port but since the data schema
is not complete yet it is not able to generate a proper file name and
save the images.