Add file name generation and test routine
The file name generation uses a self defined printer ID (like the
printer name), the portion of the PDF file name that identifies it
uniqely and the relative position (in millimeters with an accuracy of 3
decimal places).
This also adds a test routine in the main function which will be used to
create the test data.