Use AxiDraws drawing speed calculation (15%)
The AxiDraw Inkscape plugin seems to calculate the drawing speed very
different than this scripts code therefore we hardcoded the self._ms
value to what the AxiDraw code is doing (if I analysed the code
correctly).
The old 15% value resulted in a self._ms value of 3.7511134999999998.
The AxiDraws 15% value is approximately 2.41 and therefore quite a bit
slower.
We want 15% because the User Guide says it should be used for better
accuracy and we want accuracy.