Use redundancy in arrays to estimate confidence
Redundancy in this context means that there is more than one de Bruijn
window in the bit array which allows us to decode multiple positions.

This redundancy allows us to create a kind of heuristic in which the
matching positions/windows are looked up from which a crude confidence
value can be calculated.